AQLP-650 - Invalid Operator Sequence Error when I try to add Parenthesis in Custom col Formula
